refactoring, bug fix for PixelMappingModel, more tests. less of the object type checking and more of the polymorphic type checking
