+++ nexus_constructor/nexus_constructor_json/loader.py
@@ -5,21 +5,18 @@ This module contains all the parsing functions used to load the data.
 It is recommended that instead of importing this module, the root of the package be imported instead, as this exposes
 only the required root function to load the json
 """
-from nexus_constructor.data_model import (
-    Component,
-    ComponentType,
+from nexus_constructor.data_model import Component, ComponentType, Translation, Rotation
+from nexus_constructor.geometry_types import (
+    CylindricalGeometry,
+    OFFGeometry,
+    NoShapeGeometry,
+)
+from nexus_constructor.pixel_data import (
     PixelGrid,
     PixelMapping,
     SinglePixelId,
     CountDirection,
     Corner,
-    Translation,
-    Rotation,
-)
-from nexus_constructor.geometry_types import (
-    CylindricalGeometry,
-    OFFGeometry,
-    NoShapeGeometry,
 )
 from nexus_constructor.nexus import NexusDecoder
 from nexus_constructor.qml_models.instrument_model import InstrumentModel
