+++ nexus_constructor/nexus_constructor_json/writer.py
@@ -9,13 +9,11 @@ import json
 from typing import List
 from nexus_constructor.data_model import (
     Component,
-    PixelGrid,
-    PixelMapping,
-    SinglePixelId,
     Transformation,
     Translation,
     Rotation,
 )
+from nexus_constructor.pixel_data import PixelGrid, PixelMapping, SinglePixelId
 from nexus_constructor.geometry_types import OFFGeometry, Geometry, CylindricalGeometry
 from nexus_constructor.qml_models.instrument_model import InstrumentModel
