+++ nexus_constructor/nexus_filewriter_json/writer.py
@@ -8,14 +8,8 @@ only the required root function to generate the json
 Json format description can be found at https://github.com/ess-dmsc/kafka-to-nexus/
 """
 import json
-from nexus_constructor.data_model import (
-    Component,
-    Translation,
-    Rotation,
-    PixelGrid,
-    PixelMapping,
-    SinglePixelId,
-)
+from nexus_constructor.data_model import Component, Translation, Rotation
+from nexus_constructor.pixel_data import PixelGrid, PixelMapping, SinglePixelId
 from nexus_constructor.geometry_types import CylindricalGeometry, OFFGeometry
 from nexus_constructor.nexus import NexusEncoder
 from nexus_constructor.qml_models.instrument_model import InstrumentModel
