+++ nexus_constructor/qml_models/instrument_model.py
@@ -1,11 +1,5 @@
-from nexus_constructor.data_model import (
-    ComponentType,
-    PixelGrid,
-    PixelMapping,
-    Component,
-    Rotation,
-    Translation,
-)
+from nexus_constructor.data_model import ComponentType, Component, Rotation, Translation
+from nexus_constructor.pixel_data import PixelGrid, PixelMapping
 from nexus_constructor.qml_models import change_value, generate_unique_name
 from nexus_constructor.qml_models.transform_model import TransformationModel
 from nexus_constructor.off_renderer import OffMesh
